feat: Consolidate python apis in a single module 'dreal'
