+++ dreal/BUILD.bazel
@@ -1,4 +1,5 @@
 load("//third_party:com_github_robotlocomotion_drake/tools/workspace/cpplint.bzl", "cpplint")
+load("//third_party:com_github_robotlocomotion_drake/tools/workspace/python_lint.bzl", "python_lint")
 load(
     "//tools:dreal.bzl",
     "DREAL_VERSION",
@@ -6,6 +7,7 @@ load(
     "dreal_cc_binary",
     "dreal_cc_library",
     "dreal_generate_file",
+    "dreal_pybind_library",
 )
 load("@bazel_tools//tools/build_defs/pkg:pkg.bzl", "pkg_tar")
 
@@ -52,25 +54,134 @@ dreal_cc_binary(
 # ----------------------
 # Bindings
 # ----------------------
-dreal_generate_file(
-    name = "init_py",
-    out = "__init__.py",
-    content = """from dreal.symbolic import *
-from dreal.util import *
-from dreal.solver import *
-from dreal.api import *""",
+dreal_pybind_library(
+    name = "dreal_py",
+    cc_deps = [
+        "//dreal/symbolic",
+    ],
+    cc_srcs = ["dreal_py.cc"],
+    py_srcs = ["__init__.py"],
+)
+
+py_test(
+    name = "api_test_py",
+    srcs = [
+        "test/python/api_test.py",
+    ],
+    main = "test/python/api_test.py",
+    srcs_version = "PY2AND3",
+    tags = [
+        "no_asan",
+        "no_lsan",
+        "no_tsan",
+        "no_ubsan",
+    ],
+    deps = [
+        ":dreal_py",
+    ],
+)
+
+py_test(
+    name = "util_test_py",
+    srcs = [
+        "test/python/util_test.py",
+    ],
+    main = "test/python/util_test.py",
+    srcs_version = "PY2AND3",
+    tags = [
+        "no_asan",
+        "no_lsan",
+        "no_tsan",
+        "no_ubsan",
+    ],
+    deps = [
+        ":dreal_py",
+    ],
+)
+
+py_test(
+    name = "symbolic_test_py",
+    srcs = [
+        "test/python/symbolic_test.py",
+    ],
+    main = "test/python/symbolic_test.py",
+    srcs_version = "PY2AND3",
+    tags = [
+        "no_asan",
+        "no_lsan",
+        "no_tsan",
+        "no_ubsan",
+    ],
+    deps = [
+        ":dreal_py",
+    ],
+)
+
+py_test(
+    name = "solver_test_py",
+    srcs = [
+        "test/python/solver_test.py",
+    ],
+    main = "test/python/solver_test.py",
+    srcs_version = "PY2AND3",
+    tags = [
+        "no_asan",
+        "no_lsan",
+        "no_tsan",
+        "no_ubsan",
+    ],
+    deps = [
+        ":dreal_py",
+    ],
+)
+
+py_test(
+    name = "smt2_test_py",
+    srcs = [
+        "test/python/smt2_test.py",
+    ],
+    main = "test/python/smt2_test.py",
+    srcs_version = "PY2AND3",
+    tags = [
+        "no_asan",
+        "no_lsan",
+        "no_tsan",
+        "no_ubsan",
+    ],
+    deps = [
+        ":dreal_py",
+    ],
 )
 
 pkg_tar(
     name = "package_python",
     srcs = [
-        ":init_py",
+        ":_dreal_py.so",
+        ":dreal_py",
     ],
     extension = "tar.gz",
     package_dir = PYTHON_PACKAGE_DIR + "/dreal",
     tags = ["manual"],
     visibility = ["//:__pkg__"],
 )
+
+
+# dreal_generate_file(
+#     name = "init_py",
+#     out = "__init__.py",
+#     content = """from dreal.bindings.python import *""",
+# )
+
+# pkg_tar(
+#     name = "package_python",
+#     srcs = [
+#         ":init_py",
+#     ],
+#     extension = "tar.gz",
+#     package_dir = PYTHON_PACKAGE_DIR + "/dreal",
+#     tags = ["manual"],
+#     visibility = ["//:__pkg__"],
+# )
 
 # ----------------------
 # Header files to expose
@@ -119,6 +230,8 @@ pkg_tar(
 )
 
 cpplint()
+
+python_lint()
 
 licenses(["notice"])  # Apache 2.0
