+++ dreal/__init__.py
@@ -4,11 +4,11 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from dreal.symbolic._dreal_symbolic_py import *
+from dreal._dreal_py import *
 from functools import reduce
 
 # Explicitly import private symbols
-from dreal.symbolic._dreal_symbolic_py import __logical_and, __logical_or
+from dreal._dreal_py import __logical_and, __logical_or
 
 
 def logical_and(*formulas):
