+++ dreal/smt2/BUILD.bazel
@@ -2,10 +2,7 @@ load("//third_party:com_github_google_kythe/tools/build_rules/lexyacc.bzl", "gen
 load("//third_party:com_github_robotlocomotion_drake/tools/workspace/cpplint.bzl", "cpplint")
 load(
     "//tools:dreal.bzl",
-    "PYTHON_PACKAGE_DIR",
     "dreal_cc_library",
-    "dreal_generate_file",
-    "dreal_pybind_library",
 )
 load("@bazel_tools//tools/build_defs/pkg:pkg.bzl", "pkg_tar")
 
@@ -119,51 +116,6 @@ dreal_cc_library(
         "//dreal/util:scoped_unordered_map",
         "//dreal/util:string_to_interval",
     ],
-)
-
-# ----------------------
-# Bindings
-# ----------------------
-dreal_generate_file(
-    name = "init_py",
-    out = "__init__.py",
-    content = """from dreal.smt2._dreal_smt2_py import *""",
-)
-
-dreal_pybind_library(
-    name = "dreal_smt2_py",
-    cc_srcs = ["smt2_py.cc"],
-    py_srcs = ["__init__.py"],
-)
-
-py_test(
-    name = "smt2_test_py",
-    srcs = [
-        "test/smt2_test.py",
-    ],
-    main = "test/smt2_test.py",
-    srcs_version = "PY2AND3",
-    tags = [
-        "no_asan",
-        "no_lsan",
-        "no_tsan",
-        "no_ubsan",
-    ],
-    deps = [
-        ":dreal_smt2_py",
-    ],
-)
-
-pkg_tar(
-    name = "package_python",
-    srcs = [
-        ":_dreal_smt2_py.so",
-        ":dreal_smt2_py",
-    ],
-    extension = "tar.gz",
-    package_dir = PYTHON_PACKAGE_DIR + "/dreal/smt2",
-    tags = ["manual"],
-    visibility = ["//:__pkg__"],
 )
 
 # ----------------------
