+++ dreal/symbolic/BUILD.bazel
@@ -2,10 +2,8 @@ load("//third_party:com_github_robotlocomotion_drake/tools/workspace/cpplint.bzl
 load("//third_party:com_github_robotlocomotion_drake/tools/workspace/python_lint.bzl", "python_lint")
 load(
     "//tools:dreal.bzl",
-    "PYTHON_PACKAGE_DIR",
     "dreal_cc_googletest",
     "dreal_cc_library",
-    "dreal_pybind_library",
 )
 load("@bazel_tools//tools/build_defs/pkg:pkg.bzl", "pkg_tar")
 
@@ -77,48 +75,6 @@ dreal_cc_googletest(
         ":symbolic",
         ":symbolic_test_util",
     ],
-)
-
-# ----------------------
-# Bindings
-# ----------------------
-dreal_pybind_library(
-    name = "dreal_symbolic_py",
-    cc_deps = [
-        "//dreal/util:exception",
-    ],
-    cc_srcs = ["symbolic_py.cc"],
-    py_srcs = ["__init__.py"],
-)
-
-py_test(
-    name = "symbolic_test_py",
-    srcs = [
-        "test/symbolic_test.py",
-    ],
-    main = "test/symbolic_test.py",
-    srcs_version = "PY2AND3",
-    tags = [
-        "no_asan",
-        "no_lsan",
-        "no_tsan",
-        "no_ubsan",
-    ],
-    deps = [
-        ":dreal_symbolic_py",
-    ],
-)
-
-pkg_tar(
-    name = "package_python",
-    srcs = [
-        ":_dreal_symbolic_py.so",
-        ":dreal_symbolic_py",
-    ],
-    extension = "tar.gz",
-    package_dir = PYTHON_PACKAGE_DIR + "/dreal/symbolic",
-    tags = ["manual"],
-    visibility = ["//:__pkg__"],
 )
 
 # ----------------------
