+++ dreal/test/python/smt2_test.py
@@ -1,6 +1,10 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from dreal.smt2 import Logic
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from dreal import *
 
 import unittest
