+++ dreal/test/python/solver_test.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from dreal.solver import Config, Context
-from dreal.symbolic import Variable
-from dreal.util import Box
-from dreal.smt2 import Logic
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from dreal import *
 
 import unittest
 
@@ -109,7 +110,6 @@ class ContextTest(unittest.TestCase):
         # Simply check if we can do this without checking the version
         # string.
         self.assertTrue(Context.version)
-
 
 
 if __name__ == '__main__':
