+++ dreal/util/BUILD.bazel
@@ -2,11 +2,8 @@ load("//third_party:com_github_robotlocomotion_drake/tools/workspace/cpplint.bzl
 load("//third_party:com_github_robotlocomotion_drake/tools/workspace/python_lint.bzl", "python_lint")
 load(
     "//tools:dreal.bzl",
-    "PYTHON_PACKAGE_DIR",
     "dreal_cc_googletest",
     "dreal_cc_library",
-    "dreal_generate_file",
-    "dreal_pybind_library",
 )
 load("@bazel_tools//tools/build_defs/pkg:pkg.bzl", "pkg_tar")
 
@@ -440,54 +437,6 @@ dreal_cc_googletest(
     deps = [
         ":timer",
     ],
-)
-
-# ----------------------
-# Bindings
-# ----------------------
-dreal_generate_file(
-    name = "init_py",
-    out = "__init__.py",
-    content = "from dreal.util._dreal_util_py import *",
-)
-
-dreal_pybind_library(
-    name = "dreal_util_py",
-    cc_srcs = ["util_py.cc"],
-    py_deps = [
-        "//dreal/symbolic:dreal_symbolic_py",
-    ],
-    py_srcs = [":init_py"],
-)
-
-py_test(
-    name = "util_test_py",
-    srcs = [
-        "test/util_test.py",
-    ],
-    main = "test/util_test.py",
-    srcs_version = "PY2AND3",
-    tags = [
-        "no_asan",
-        "no_lsan",
-        "no_tsan",
-        "no_ubsan",
-    ],
-    deps = [
-        ":dreal_util_py",
-    ],
-)
-
-pkg_tar(
-    name = "package_python",
-    srcs = [
-        ":_dreal_util_py.so",
-        ":dreal_util_py",
-    ],
-    extension = "tar.gz",
-    package_dir = PYTHON_PACKAGE_DIR + "/dreal/util",
-    tags = ["manual"],
-    visibility = ["//:__pkg__"],
 )
 
 # ----------------------
